* echo.cpp: add var handle
* Kernel.userprefs:
* exit.cpp:
* help.cpp:
* lsdev.cpp: .

* set.cpp: varibale handle

* Kernel.cproj: ..

* Makefile: add new

* config.h: Update version

* mm.h: remove PAGE SIZE

* split.hpp: ADD New Version

* string.hpp: Nothing

* shell.hpp: Add var handle

* font12x16.c: remove

* mm.c: change uint64_t -> uint32_t

* main.cpp: nothing

* split.cpp: add new version

* shell.cpp: varibale

+++ inc/kernel/mm.h
@@ -8,7 +8,7 @@ extern "C"
 
 #include <types.h>
 
-int 				mm_init(uint64_t);
+int 				mm_init(uint32_t);
 void 				mm_test();
 void 				mm_free(void *);
 intptr_t*			mm_alloc(size_t);
